% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define-priors.R
\name{default_priors}
\alias{default_priors}
\title{Define default prior functions for selection model parameters}
\usage{
default_priors(
  beta_mean = 0,
  beta_sd = 1,
  tau_mode = 0.2,
  tau_alpha = 1,
  lambda_mode = 0.8,
  lambda_alpha = 1
)
}
\arguments{
\item{beta_mean}{numeric vector of prior means for beta (mean regression)
parameters.}

\item{beta_sd}{numeric vector of prior standard deviations for beta (mean
regression) parameters.}

\item{tau_mode}{numeric vector of prior modes for tau (heterogeneity SD)
regression parameters.}

\item{tau_alpha}{numeric vector of prior precisions for tau (heterogeneity SD)
regression parameters.}

\item{lambda_mode}{numeric vector of prior modes for lambda (selection)
parameters.}

\item{lambda_alpha}{numeric vector of prior precisions for lambda
(selection) parameters.}
}
\value{
An object of class \code{"selmodel_prior"} containing the following
components:
\describe{
\item{\code{log_prior}}{A function with arguments \code{beta},\code{gamma},\code{zeta} that returns the log of the prior density over these parameters.}
\item{\code{score_prior}}{A function with arguments \code{beta},\code{gamma},\code{zeta} that returns the vector of scores for the prior density over these parameters.}
\item{\code{hessian_prior}}{A function with arguments \code{beta},\code{gamma},\code{zeta} that returns the Hessian matrix of the prior density over these parameters.}
}
}
\description{
Creates a set of priors for use in estimating selection models.
beta parameters are assigned independent normal priors. log(tau) parameters
are assigned independent log-gamma priors. log(lambda) parameters are
assigned independent log-gamma priors.
}
\examples{
# set very informative priors on beta and lambda
strong_priors <- default_priors(beta_mean = 0.4, beta_sd = 0.1, lambda_mode = 0.2, lambda_sd = 0.1)

# set very weak priors
weak_priors <- default_priors(beta_sd = 10, tau_sd = 1, lambda_sd = 4)

}
