% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing-effects.R
\name{missing_effects}
\alias{missing_effects}
\title{Estimate number of missing effect sizes}
\usage{
missing_effects(object)
}
\arguments{
\item{object}{Fitted model of class inheriting from \code{"selmodel"}.}
}
\description{
Estimate number of missing effect sizes based on a fitted \code{selmodel} object.
}
\examples{

res <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  selection_type = "step",
  steps = 0.025,
  mean_mods = ~ 0 + sample_population,
  var_mods = ~ 0 + sample_population
)

missing_effects(res)

}
