% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{n_ES_param}
\alias{n_ES_param}
\title{Simulate empirical distribution of sample size and number of effect
sizes}
\usage{
n_ES_param(mean_N, mean_ES, min_N = 20L)
}
\arguments{
\item{mean_N}{numeric value specifying the average sample size per primary study.}

\item{mean_ES}{numeric value specifying the average number of effect sizes per
primary study.}

\item{min_N}{numeric value specifying the minimum sample size per study.}
}
\value{
A function that generates a \code{data.frame} with randomly generated sample
size per primary study and number of effect sizes per study.
}
\description{
A functional that takes in average sample size per primary
study, average number of effect sizes per study, and the minimum sample
size per study and returns a function to generate random samples of primary
study sample sizes and numbers of effect sizes per study.
}
\examples{
study_features <- n_ES_param(mean_N = 40, mean_ES = 3, min_N = 10)
study_features(m = 3)
study_features(m = 5)

}
