% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.selmodel.R
\name{predict.step.selmodel}
\alias{predict.step.selmodel}
\title{Predict method for fitted \code{step.selmodel}}
\usage{
\method{predict}{step.selmodel}(
  object,
  newdata = NULL,
  location = TRUE,
  scale = TRUE,
  selection = TRUE,
  obs_prob = TRUE,
  ...
)
}
\arguments{
\item{object}{Fitted model of class inheriting from \code{"step.selmodel"}.}

\item{newdata}{Optionally, a data frame containing observations for which to
generate predictions. If \code{NULL} (the default), predictions are
generated for the data frame used to fit the model.}

\item{location}{logical with \code{TRUE} (the default) indicating that the
predicted value of the mean effect size should be returned.}

\item{scale}{logical with \code{TRUE} (the default) indicating that the predicted
value of the heterogeneity should be returned.}

\item{selection}{logical with \code{TRUE} (the default) indicating that the
predicted selection parameters should be returned.}

\item{obs_prob}{logical with \code{TRUE} (the default) indicating that the
predicted probability of observing an effect size should be returned.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Calculate predicted values for the mean (location),
heterogeneity (scale), and selection parameters from a fitted model of
class \code{"step.selmodel"}.
}
\examples{
res <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  selection_type = "step",
  steps = 0.025,
  mean_mods = ~ 0 + sample_population,
  var_mods = ~ 0 + sample_population
)

predict(res)

newdat <- data.frame(sample_population = c("Students","Students"))
predict(res, newdata = newdat, obs_prob = FALSE)

newdat$sei <- c(0.10, 0.15)
predict(res, newdata = newdat)
}
