% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_plot.R
\name{selection_plot}
\alias{selection_plot}
\alias{selection_plot.selmodel}
\alias{selection_plot.boot.selmodel}
\title{Plot the selection weights implied by an estimated selection model.}
\usage{
selection_plot(
  mod,
  limits = c(0, 1),
  pts = 200L,
  ref_pval = NULL,
  transform = "identity",
  expand = ggplot2::expansion(0, 0.01),
  ...
)

\method{selection_plot}{selmodel}(
  mod,
  limits = c(0, 1),
  pts = 200L,
  ref_pval = NULL,
  transform = "identity",
  expand = ggplot2::expansion(0, 0.01),
  fill = "blue",
  alpha = 0.5,
  step_linetype = "dashed",
  ...
)

\method{selection_plot}{boot.selmodel}(
  mod,
  limits = c(0, 1),
  pts = 200L,
  ref_pval = NULL,
  transform = "identity",
  expand = ggplot2::expansion(0, 0.01),
  color = "black",
  linewidth = 1.2,
  step_linetype = "dashed",
  draw_boots = TRUE,
  fill = "blue",
  alpha = 0.5,
  boot_color = "blue",
  boot_alpha = 0.1,
  ...
)
}
\arguments{
\item{mod}{fitted model of class \code{"selmodel"}.}

\item{limits}{numeric vector of length 2 specifying the minimum and maximum p-values to plot.}

\item{pts}{number of points for which to calculate selection weights, with a
default of 200 points, evenly spaced between the specified limits.}

\item{ref_pval}{numeric value of a p-value at which to standardize the
weights. If not \code{NULL}, then a p-value of \code{ref_pval} will have
selection weight of 1 and selection weights for all other p-values will be
calculated relative to \code{ref_pval}.}

\item{transform}{character string specifying the name of a transformation function or the transformation function itself, as defined in the scales package. The transform is passed to \code{ggplot2::scale_x_continuous}. The default transform is \code{"identity"}. Other useful transforms for p-values are \code{"sqrt"} for square-root or \code{"asn"} for the arc-sin square root.}

\item{expand}{passed to the \code{expand} argument of \code{ggplot2::scale_x_continuous}.}

\item{...}{further arguments passed to \code{ggplot2::scale_x_continuous}.}

\item{fill}{character string specifying the fill-color to use when \code{mod}
does not include bootstrap replications, with a default of \code{"blue"}.
Passed to \code{ggplot2::geom_area()}.}

\item{alpha}{numeric value specifying the opacity of the filled area plot,
with a default of 0.5. Passed to \code{ggplot2::geom_area()}. Only used
when \code{mod} does not include bootstrap replications.}

\item{step_linetype}{character string specifying the type of line to draw to
indicate p-value thresholds assumed in \code{mod}.}

\item{color}{character string specifying the line color to use for drawing
the estimated selection weights, with a default of \code{"black"}. Passed
to \code{ggplot2::geom_line()}. Only used when \code{mod} includes
bootstrap replications.}

\item{linewidth}{numeric value specifying the line width to use for drawing
the estimated selection weights, with a default of 1.2. Passed to
\code{ggplot2::geom_line()}. Only used when \code{mod} includes bootstrap
replications.}

\item{draw_boots}{logical value indicating whether to draw the selection
weights for each bootstrap replication, with a default of \code{TRUE}.}

\item{boot_color}{character string specifying the line color to use for
drawing the selection weights of each bootstrap replication, with a default
of \code{"blue"}. Passed to \code{ggplot2::geom_line()}. Only used when
\code{mod} includes bootstrap replications.}

\item{boot_alpha}{numeric value specifying the opacity of the lines for
drawing the selection weights of each bootstrap replication, with a default
of \code{"blue"}. Passed to \code{ggplot2::geom_line()}. Only used when
\code{mod} includes bootstrap replications.}
}
\value{
A \code{ggplot2} object.
}
\description{
For a fitted model of class \code{"selmodel"}, create a plot of
the selection weights implied by the model parameter estimates. If the
model includes bootstrapped confidence intervals, then the plot will also
display the selection weights implied by each bootstrap replicate of the
parameter estimates.
}
\examples{
mod <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ARGL",
  bootstrap = "none"
)

selection_plot(mod, fill = "purple")

# rescale the horizontal axis using arc-sin square root
selection_plot(mod, fill = "purple", transform = "asn") 


mod_boot <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ARGL",
  bootstrap = "multinomial",
  CI_type = "percentile",
  R = 9L
)

 selection_plot(mod_boot, transform = "sqrt")
 selection_plot(mod_boot, transform = "sqrt", draw_boots = FALSE) # turn off bootstrap lines
 selection_plot(mod_boot, transform = "sqrt", color = "red", boot_color = "orange") # change colors

}
