% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_plot.R
\name{selection_wts}
\alias{selection_wts}
\alias{selection_wts.step.selmodel}
\alias{selection_wts.beta.selmodel}
\title{Calculate model-implied weights for specified p-values.}
\usage{
selection_wts(mod, pvals, ref_pval, ...)

\method{selection_wts}{step.selmodel}(mod, pvals = NULL, ref_pval = NULL, bootstraps = TRUE, ...)

\method{selection_wts}{beta.selmodel}(mod, pvals = NULL, ref_pval = NULL, bootstraps = TRUE, ...)
}
\arguments{
\item{mod}{fitted model of class \code{"selmodel"}.}

\item{pvals}{numeric vector of p-values for which to calculate selection
weights.}

\item{ref_pval}{numeric value of a p-value at which to standardize the
weights. If not \code{NULL}, then a p-value of \code{ref_pval} will have
selection weight of 1 and selection weights for all other p-values will be
calculated relative to \code{ref_pval}.}

\item{...}{further arguments passed to some methods.}

\item{bootstraps}{If \code{mod} includes bootstrap replications, then setting
\code{bootstraps = TRUE} will return selection weights for each bootstrap
replication, in addition to the selection weights implied by the model
parameter estimates. Ignored if \code{mod} does not include bootstrap
replications.}
}
\value{
If \code{mod} does not include bootstrapped confidence intervals or
if the argument \code{bootstraps = FALSE}, then \code{selection_wts} will
return a \code{data.frame} containing the user-specified p-values and the
selection weights implied by the estimated model parameters.

If \code{mod} does include bootstrapped confidence intervals (i.e., when
\code{inherits(mod, "boot.selmodel")} is \code{TRUE}) and the argument
\code{bootstraps = TRUE}, then \code{selection_wts} will return a list with
two elements. The first element is a \code{data.frame} containing the
user-specified p-values and the selection weights implied by the estimated
model parameters. The second element is a \code{data.frame} containing the
user-specified p-values and the selection weights implied by each bootstrap
replicate of the model parameter estimates. The \code{data.frame} includes
an additional variable, \code{rep}, identifying the bootstrap replicate.
}
\description{
Calculates the selection weights implied by an estimated model
for a user-specified p-value or set of p-values.
}
\examples{
mod <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ARGL"
)

selection_wts(mod, pvals = seq(0, 1, 0.2))

mod_boot <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ARGL",
  bootstrap = "multinomial",
  CI_type = "percentile",
  R = 9
)

selection_wts(mod_boot, pvals = seq(0, 1, 0.2))


}
