% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self-control.R
\docType{data}
\name{self_control}
\alias{self_control}
\title{Self-Control Training Meta-Analysis}
\format{
A data frame with 158 rows and 26 variables:
\describe{
\item{studyid}{identifier for the study.}
\item{esid}{identifier for the effect size.}
\item{name}{name of the study.}
\item{g}{effect size in form of Hedges' g.}
\item{var_g}{corresponding variance of the effect size.}
\item{se_g}{corresponding standard error of the effect size.}
\item{outcome}{name of the outcome measure.}
\item{comparison}{conditions compared.}
\item{type_of_treatment}{type of treatment/training.}
\item{length_of_treatment}{length of treatment coded in days.}
\item{publication_status}{indicator for whether the study was published.}
\item{publication_status_new}{indicator for whether the study was published (newer version).}
\item{publication_year}{year that the study was published.}
\item{research_group}{indicator for whether researchers of the study belonged to the strength model group.}
\item{control_group_quality}{quality of the control group with active indicating that the control group worked on some task.}
\item{gender_ratio}{percentage of males in the sample.}
\item{type_of_outcome}{type of outcome.}
\item{subjectivity_of_outcome_measurement}{indicator for the subjectivity of the outcome measure.}
\item{lab_based_versus_real_world_behavior}{indicator for whether the behavior measured was assessed in lab or in the real-world.}
\item{stamina_versus_strength}{indicator for whether the outcome was assessed with (Stamina) or without a preceding effortful task (Strength).}
\item{pre_test_measurement}{indicator for whether there was a pre and post measure of outcome or only post measure.}
\item{sample_population}{population examined in the study.}
\item{sample_age}{average age of the sample.}
\item{attrition}{percentage of attrition.}
\item{participant_compensation}{type of compensation received by the participants.}
\item{self_control_potential}{indicator for whether the outcome measure required utilization of maximum self-control potential.}
}
}
\source{
\href{https://osf.io/v7gxf/}{OSF page for the project}
}
\usage{
self_control
}
\description{
Meta-analytic dataset containing results of primary studies examining the
effect of self-control training.
}
\references{
\insertRef{friese2017}{metaselection}
}
\keyword{datasets}
