% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwc_es.R
\docType{data}
\name{wwc_es}
\alias{wwc_es}
\title{What Works Clearinghouse sample size and effect size distribution data}
\format{
A tibble with 615 rows and 2 variables:
\describe{
\item{n}{the sample size of the primary study.}
\item{n_ES}{number of effect sizes in the primary study.}
}
}
\usage{
wwc_es
}
\description{
A dataset containing sample size and number of effect sizes seen in primary
studies evaluated by the What Works Clearinghouse.
}
\references{
\insertRef{wwc}{metaselection}
}
\keyword{datasets}
