% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{beta_fun}
\alias{beta_fun}
\title{Censor meta-analytic dataset based on the beta-density model model}
\usage{
beta_fun(delta_1 = 1, delta_2 = 1, trunc_1 = 0.025, trunc_2 = 0.975)
}
\arguments{
\item{delta_1}{a number}

\item{delta_2}{a number}

\item{trunc_1}{a number}

\item{trunc_2}{a number}
}
\value{
A function that can be used to censor a meta-analytic dataset based on the beta-density model.
}
\description{
A functional that takes in beta-density model parameters and returns a function that can be used to censor meta-analytic datasets according to the beta-density model.
}
