% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaselection-package.R
\docType{package}
\name{metaselection-package}
\alias{metaselection}
\alias{metaselection-package}
\title{metaselection: Meta-Analytic Selection Models with Cluster-Robust and Cluster-Bootstrap Standard Errors for Dependent Effect Size Estimates}
\description{
Fits a flexible class of p-value selection models for meta-analysis and meta-regression models, providing standard errors and confidence intervals based on either cluster-robust variance estimators (i.e., sandwich estimators) or cluster-level bootstrapping to handle dependent effect size estimates, as described in Pustejovsky, Citkowicz, and Joshi (2025) \doi{10.31222/osf.io/qg5x6_v1} and Citkowicz, Pustejovsky, and Joshi (2026) \doi{10.31222/osf.io/wjpxk_v1}. Supported models include generalizations of the step-function selection model as proposed by Vevea and Hedges (1995) \doi{10.1007/BF02294384} and the beta-function selection model as proposed by Citkowicz and Vevea (2017) \doi{10.1037/met0000119}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jepusto/metaselection}
}

}
\author{
\strong{Maintainer}: James E. Pustejovsky \email{jepusto@gmail.com} (\href{https://orcid.org/0000-0003-0591-9465}{ORCID})

Authors:
\itemize{
  \item James E. Pustejovsky \email{jepusto@gmail.com} (\href{https://orcid.org/0000-0003-0591-9465}{ORCID})
  \item Megha Joshi \email{megha.j456@gmail.com} (\href{https://orcid.org/0000-0001-7936-076X}{ORCID})
  \item Martyna Citkowicz \email{martyna.citkowicz@gmail.com} (\href{https://orcid.org/0000-0002-2048-0625}{ORCID})
}

}
\keyword{internal}
