% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaselection-package.R
\docType{package}
\name{metaselection-package}
\alias{metaselection}
\alias{metaselection-package}
\title{metaselection: Meta-analytic selection models with cluster-robust and cluster-bootstrap standard errors}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: James Pustejovsky \email{jepusto@gmail.com} (\href{https://orcid.org/0000-0003-0591-9465}{ORCID})

Authors:
\itemize{
  \item Megha Joshi \email{megha.j456@gmail.com} (\href{https://orcid.org/0000-0001-7936-076X}{ORCID})
}

}
\keyword{internal}
