% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{n_ES_param}
\alias{n_ES_param}
\title{Simulate empirical distribution of sample size and number of effect sizes}
\usage{
n_ES_param(mean_N, mean_ES, min_N = 20L)
}
\arguments{
\item{mean_N}{number indicating the average sample size per primary study}

\item{mean_ES}{number indicating the average number of effect sizes per primary study}

\item{min_N}{number indicating the minimum sample size per study}
}
\value{
A function that generates a dataframe with randomly generated sample size per primary study and number of effect sizes per study
}
\description{
A functional that takes in average sample size per primary study, average number of effect sizes per study, and the minimum sample size per study and returns a function to generate dataset with random samples of primary study sample size and and the number of effect sizes per study based on the Poisson distribution.
}
