% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practice-facilitation.R
\docType{data}
\name{practice_facilitation}
\alias{practice_facilitation}
\title{Practice Facilitation Meta-Analysis}
\format{
A data frame with 23 rows and 3 variables:
\describe{
\item{author}{first author and publication year of primary study report.}
\item{score}{score on a scale from 0 to 12, for which higher scores correspond to higher quality of the study methods.}
\item{design}{study design, with CCT = controlled clinical trial, C-RCT = cluster randomized controlled trial, RCT = randomized controlled trial.}
\item{allocation_concealed}{indicator for allocation concealment.}
\item{blinded}{indicator for whether study was single- or double-blinded.}
\item{intent_to_treat}{indicator for whether study adhered to intent-to-treat principle.}
\item{outcome}{description of outcome measure.}
\item{follow_up}{months of follow-up.}
\item{retention_pct}{percentage of sample retained at follow-up.}
\item{SMD}{effect size in form of Hedges' g.}
\item{SE}{corresponding variance of the effect size.}
}
}
\source{
Table 1 of Baskerville et al. (2012; \doi{10.1370/afm.1312}).
}
\usage{
practice_facilitation
}
\description{
Meta-analytic dataset containing results of primary studies examining the
effect of practice facilitation on the uptake of evidence-based practices
(EBPs) in primary care settings.
}
\references{
\insertRef{baskerville2012}{metaselection}
}
\keyword{datasets}
