% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.selmodel.R
\name{print.selmodel}
\alias{print.selmodel}
\title{Print results from \code{selmodel()}}
\usage{
\method{print}{selmodel}(x, transf_gamma = FALSE, transf_zeta = FALSE, ...)
}
\arguments{
\item{x}{Fitted model of class \code{"selmodel"}.}

\item{transf_gamma}{logical with \code{TRUE} indicating that the heterogeneity parameter estimates (called gamma) should be transformed by exponentiating.}

\item{transf_zeta}{logical with \code{TRUE} indicating that the selection parameter estimates (called zeta) should be transformed by exponentiating.}

\item{...}{Not used.}
}
\description{
Print relevant results from \code{selmodel()}
}
