% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.selmodel.R
\name{print.selmodel}
\alias{print.selmodel}
\title{Print results from \code{selmodel()}}
\usage{
\method{print}{selmodel}(x, exponentiate = TRUE)
}
\arguments{
\item{x}{Fitted model of class \code{"selmodel"}.}

\item{exponentiate}{logical with \code{TRUE} indicating that the gamma and zeta estimates should be exponentiated}
}
\description{
Print relevant results from \code{selmodel()}
}
