% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_model.R
\name{selection_model}
\alias{selection_model}
\title{Estimate step or beta selection model}
\usage{
selection_model(
  data,
  yi,
  vi,
  sei,
  ai,
  cluster,
  selection_type = c("step", "beta"),
  alternative = "greater",
  steps = NULL,
  mean_mods = NULL,
  var_mods = NULL,
  sel_mods = NULL,
  sel_zero_mods = NULL,
  priors = define_priors(),
  subset = NULL,
  estimator = "CML",
  vcov_type = "robust",
  CI_type = "large-sample",
  conf_level = 0.95,
  theta = NULL,
  optimizer = NULL,
  optimizer_control = list(),
  use_jac = NULL,
  bootstrap = "none",
  R = 1999,
  retry_bootstrap = 0L,
  valence_check = TRUE,
  ...
)
}
\arguments{
\item{data}{\code{data.frame} or \code{tibble} containing the
meta-analytic data.}

\item{yi}{vector of effect sizes estimates.}

\item{vi}{vector of sample variances. If \code{vi} is
specified, then the \code{sei} argument must be omitted.}

\item{sei}{vector of sampling standard errors. If \code{sei}
is specified, then the \code{vi} argument must be omitted.}

\item{ai}{optional vector of analytic weights.}

\item{cluster}{vector indicating which observations belong to
the same cluster.}

\item{selection_type}{character string specifying the type
selection model to estimate, with possible options
\code{"step"} or \code{"beta"}.}

\item{alternative}{character string specifying the direction
of the alternative hypothesis used in computing p-values
for the observed effect sizes, with possible options
\code{"greater"} (the default) or \code{"less"}.}

\item{steps}{If \code{selection_type = "step"}, a numeric
vector of one or more values specifying the thresholds (or
steps) where the selection probability changes, with a
default of \code{steps = .025}. If \code{selection_type =
"beta"}, then a numeric vector of two values specifying the
thresholds beyond which the selection function is
truncated, with a default of \code{steps = c(.025, .975)}.}

\item{mean_mods}{optional model formula for moderators
related to average effect size magnitude.}

\item{var_mods}{optional model formula for moderators related
to effect size heterogeneity.}

\item{sel_mods}{optional model formula for moderators related
to the probability of selection. Only relevant for
\code{selection_type = "step"}.}

\item{sel_zero_mods}{optional model formula for moderators
related to the probability of selection for p-values below
the lowest threshold value of \code{steps}. Only relevant
for \code{selection_type = "step"}.}

\item{priors}{a \code{selmodel_prior} object that defines
priors (i.e., penalty terms) for model parameters, with a
default of \code{define_priors()}. Set to \code{NULL} to
obtain unpenalized estimates.}

\item{subset}{optional logical expression indicating a subset
of observations to use for estimation.}

\item{estimator}{vector indicating whether to use the
composite marginal likelihood estimator (option
\code{"CML"}) or the augmented and reweighted Gaussian
likelihood estimator (option \code{"ARGL"} or
\code{"ARGL-full"}). If \code{selection_type = "beta"},
only the composite marginal likelihood estimator,
\code{"CML"}, is available. For step function models, both
estimators are available.}

\item{vcov_type}{character string specifying the type of
variance-covariance matrix to calculate, with possible
options \code{"robust"} for robust or cluster-robust
standard errors, \code{"model-based"} for model-based
standard errors, or \code{"none"}.}

\item{CI_type}{character string or vector specifying the type
of confidence interval to calculate, with possible options
\code{"large-sample"} for large-sample normal interval (the
default), \code{"percentile"} for a percentile interval,
\code{"BCa"} for a bias-corrected-and-accelerated interval,
\code{"bias-corrected"} for a bias-corrected percentile
interval (without acceleration correction), \code{"normal"}
for a standard normal interval, \code{"basic"} for a basic
interval, \code{"student"} for a studentized interval, or
\code{"none"}. More than one type of interval can be
computed by specifying a character vector with multiple
options.}

\item{conf_level}{desired coverage level for confidence
intervals, with the default value set to \code{.95}.}

\item{theta}{optional numeric vector of starting values to
use in optimization routines.}

\item{optimizer}{character string indicating the optimizer to
use. Ignored if \code{estimator = "ARGL"} or
\code{"ARGL-full"}.}

\item{optimizer_control}{an optional list of control
parameters to be used for optimization}

\item{use_jac}{logical indicating whether to use the Jacobian
of the estimating equations for optimization. If
\code{NULL} (the default), it will be reset to \code{FALSE}
if \code{estimator = "CML"} or to \code{TRUE} if
\code{estimator = "ARGL"}}

\item{bootstrap}{character string specifying the type of
bootstrap to run, with possible options \code{"none"} (the
default), \code{"exponential"} for the fractionally
re-weighted cluster bootstrap, \code{"multinomial"} for a
conventional clustered bootstrap, or , \code{"two-stage"}
for a two-stage clustered bootstrap.}

\item{R}{number of bootstrap replications, with a default of
\code{1999}.}

\item{retry_bootstrap}{number of times to re-draw a bootstrap
sample in the event of non-convergence, with a default of
\code{0}.}

\item{valence_check}{logical value controlling whether to
check that the valence of the median effect size estimate
is consistent with the direction of the specified
\code{alternative}. If \code{TRUE} (the default), a warning
will be issued when most effect size estimates have the
opposite sign of \code{alternative}. Set to \code{FALSE} to
suppress the warning.}

\item{...}{further arguments passed to
\code{simhelpers::bootstrap_CIs}.}
}
\value{
An object of class \code{"selmodel"} containing the
following components:
\describe{
\item{\code{est}}{\code{data.frame} with parameter estimates, standard errors, and
confidence intervals. Note that the results do not include p-values so
as to focus interpretation on the parameter estimates, rather than on
the statistical significance of any given parameter.}
\item{\code{vcov}}{matrix containing the estimated variance-covariance matrix
of the parameter estimates.}
\item{\code{method}}{character string indicating the optimization method used to solve for parameter estimates.}
\item{\code{info}}{further information about the optimization results.}
\item{\code{ll}}{log likelihood of the model evaluated at the reported parameter estimates.}
\item{\code{wpll}}{weighted partial log likelihood of the random effects model, with weights corresponding to inverse selection probabilities.}
\item{\code{n_clusters}}{number of independent clusters of effect sizes.}
\item{\code{n_effects}}{number of effect size estimates in the data.}
\item{\code{...}}{some additional elements containing information about the methods used to estimate the model.}
}
}
\description{
Estimate step or beta selection model, with
standard errors and confidence intervals based on either
cluster-robust variance estimators (i.e., sandwich
estimators) or cluster-level bootstrapping to handle
dependent effect size estimates.
}
\examples{
res_ML <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = 0.025,
  estimator = "CML",
  bootstrap = "none"
)

res_ML
summary(res_ML)

# configure progress bar
progressr::handlers(global = TRUE)

res_hybrid <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = 0.025,
  estimator = "ARGL",
  bootstrap = "multinomial",
  CI_type = "percentile",
  R = 19
)

res_hybrid
summary(res_hybrid)

}
