% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_plot.R
\name{selection_plot}
\alias{selection_plot}
\alias{selection_plot.selmodel}
\alias{selection_plot.boot.selmodel}
\title{Plot the selection weights implied by an estimated selection model.}
\usage{
selection_plot(mod, pts = 200L, ref_pval = NULL, ...)

\method{selection_plot}{selmodel}(
  mod,
  pts = 200L,
  ref_pval = NULL,
  fill = "blue",
  alpha = 0.5,
  step_linetype = "dashed",
  ...
)

\method{selection_plot}{boot.selmodel}(
  mod,
  pts = 200L,
  ref_pval = NULL,
  color = "black",
  linewidth = 1.2,
  step_linetype = "dashed",
  draw_boots = TRUE,
  boot_color = "blue",
  boot_alpha = 0.1,
  ...
)
}
\arguments{
\item{mod}{Fitted model of class \code{"selmodel"}.}

\item{pts}{Number of points for which to calculate selection weights, with a
default of 200 points, evenly spaced between 0 and 1.}

\item{ref_pval}{Numeric value of a p-value at which to standardize the
weights. If not \code{NULL}, then a p-value of \code{ref_pval} will have
selection weight of 1 and selection weights for all other p-values will be
calculated relative to \code{ref_pval}.}

\item{...}{further arguments passed to some methods.}

\item{fill}{character string specifying the fill-color to use when \code{mod}
does not include bootstrap replications, with a default of \code{"blue"}.
Passed to \code{ggplot2::geom_area()}.}

\item{alpha}{numeric value specifying the opacity of the filled area plot,
with a default of 0.5. Passed to \code{ggplot2::geom_area()}. Only used
when \code{mod} does not include bootstrap replications.}

\item{step_linetype}{character string specifying the type of line to draw to
indicate p-value thresholds assumed in \code{mod}.}

\item{color}{character string specifying the line color to use for drawing
the estimated selection weights, with a default of \code{"black"}. Passed
to \code{ggplot2::geom_line()}. Only used when \code{mod} includes
bootstrap replications.}

\item{linewidth}{numeric value specifying the line width to use for drawing
the estimated selection weights, with a default of 1.2. Passed to
\code{ggplot2::geom_line()}. Only used when \code{mod} includes bootstrap
replications.}

\item{draw_boots}{logical value indicating whether to draw the selection
weights for each bootstrap replication, with a default of \code{TRUE}.}

\item{boot_color}{character string specifying the line color to use for
drawing the selection weights of each bootstrap replication, with a default
of \code{"blue"}. Passed to \code{ggplot2::geom_line()}. Only used when
\code{mod} includes bootstrap replications.}

\item{boot_alpha}{numeric value specifying the opacity of the lines for
drawing the selection weights of each bootstrap replication, with a default
of \code{"blue"}. Passed to \code{ggplot2::geom_line()}. Only used when
\code{mod} includes bootstrap replications.}
}
\value{
A \code{ggplot2} object.
}
\description{
For a fitted model of class \code{"selmodel"}, create a plot of
the selection weights implied by the model parameter estimates. If the
model includes bootstrapped confidence intervals, then the plot will also
display the selection weights implied by each bootstrap replicate of the
parameter estimates.
}
\examples{
mod <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ML",
  bootstrap = "none"
)

selection_plot(mod, fill = "purple")


mod_boot <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = c(0.025, .5),
  estimator = "ML",
  bootstrap = "multinomial",
  CI_type = "percentile",
  R = 9
)

 selection_plot(mod_boot)
 selection_plot(mod_boot, draw_boots = FALSE) # turn off bootstrap lines
 selection_plot(mod_boot, color = "red", boot_color = "orange") # change colors

}
