% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-generation.R
\name{step_fun}
\alias{step_fun}
\title{Censor meta-analytic dataset based on a step-function model}
\usage{
step_fun(cut_vals = 0.025, weights = 1, renormalize = TRUE)
}
\arguments{
\item{cut_vals}{numeric vector of one or more values specifying the thresholds of p-values where the likelihood of effects being selected changes}

\item{weights}{numeric vector of one or more values specifying the likelihood of selection for different intervals of p-values; the intervals are determined by the \code{cut_vals}}

\item{renormalize}{logical with \code{TRUE} indicating}
}
\value{
A function that can be used to censor a meta-analytic dataset based on the step-function model.
}
\description{
A functional that takes in cut values and weights representing selection probabilities for different intervals of p-values and returns a function that can be used to censor meta-analytic datasets according to the step-function model.
}
