% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.selmodel.R
\name{summary.selmodel}
\alias{summary.selmodel}
\title{Summarize results from a \code{selmodel} object}
\usage{
\method{summary}{selmodel}(object, transf_gamma = TRUE, transf_zeta = TRUE, digits = 3, ...)
}
\arguments{
\item{object}{fitted model of class \code{"selmodel"}.}

\item{transf_gamma}{logical with \code{TRUE} (the default) indicating that the heterogeneity parameter estimates (called gamma) should be transformed by exponentiating.}

\item{transf_zeta}{logical with \code{TRUE} (the default) indicating that the selection parameter estimates (called zeta) should be transformed by exponentiating.}

\item{digits}{minimum number of significant digits to be used, with a default of 3.}

\item{...}{further arguments passed to \code{print.data.frame()}.}
}
\description{
Summarize relevant results from a fitted selection model.
}
\examples{
res_ML <- selection_model(
  data = self_control,
  yi = g,
  sei = se_g,
  cluster = studyid,
  steps = 0.025,
  estimator = "CML",
  bootstrap = "none"
)

summary(res_ML)
summary(res_ML, transf_gamma = FALSE, transf_zeta = FALSE)
}
