% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.selmodel.R
\name{summary.selmodel}
\alias{summary.selmodel}
\title{Summarize results from \code{selmodel()}}
\usage{
\method{summary}{selmodel}(x, digits = 3, transf_gamma = FALSE, transf_zeta = FALSE, ...)
}
\arguments{
\item{x}{Fitted model of class \code{"selmodel"}.}

\item{transf_gamma}{logical with \code{TRUE} indicating that the heterogeneity parameter estimates (called gamma) should be transformed by exponentiating.}

\item{transf_zeta}{logical with \code{TRUE} indicating that the selection parameter estimates (called zeta) should be transformed by exponentiating.}

\item{...}{Not used.}
}
\description{
Summarize relevant results from \code{selmodel()}
}
