% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwc_es.R
\docType{data}
\name{wwc_es}
\alias{wwc_es}
\title{What Works Clearinghouse sample size and effect size distribution data}
\format{
A tibble with 625 rows and 2 variables:
\describe{
\item{n}{the number of studies included in the meta-analysis.}
\item{n_ES}{correlation between outcomes.}
}
}
\usage{
wwc_es
}
\description{
A dataset containing sample size and number of effect sizes seen in primary
studies evaluated by the What Works Clearinghouse.
}
\keyword{datasets}
